+++ com.unity.render-pipelines.high-definition/Editor/Material/Lit/ShaderGraph/HDLitSubShader.cs
@@ -359,17 +359,6 @@ namespace UnityEditor.Rendering.HighDefinition
             {
                 "#include \"Packages/com.unity.render-pipelines.high-definition/Runtime/RenderPipeline/ShaderPass/ShaderPassDistortion.hlsl\"",
             },
-            StencilOverride = new List<string>()
-            {
-                "// Stencil setup",
-                "Stencil",
-                "{",
-                string.Format("   WriteMask {0}", (int)StencilUsage.DistortionVectors),
-                string.Format("   Ref  {0}", (int)StencilUsage.DistortionVectors),
-                "   Comp Always",
-                "   Pass Replace",
-                "}"
-            },
             PixelShaderSlots = new List<int>()
             {
                 HDLitMasterNode.AlphaSlotId,
@@ -634,7 +623,42 @@ namespace UnityEditor.Rendering.HighDefinition
                 HDLitMasterNode.VertexNormalSlotID,
                 HDLitMasterNode.VertexTangentSlotID
             },
-            UseInPreview = true,
+            UseInPreview = false,
+        };
+
+        Pass m_PassRayTracingPrepass = new Pass()
+        {
+            Name = "RayTracingPrepass",
+            LightMode = "RayTracingPrepass",
+            TemplateName = "HDLitPass.template",
+            MaterialName = "Lit",
+            ShaderPassName = "SHADERPASS_DEPTH_ONLY",
+            BlendOverride = "Blend One Zero",
+            ZWriteOverride = "ZWrite On",
+            CullOverride = HDSubShaderUtilities.defaultCullMode,
+            ShaderStages = HDSubShaderUtilities.s_ShaderStagesRasterization,
+            ExtraDefines = new List<string>()
+            {
+                "#define RAYTRACINGPREPASS",
+                DefineRaytracingKeyword(RayTracingNode.RaytracingVariant.High),
+            },
+            Includes = new List<string>()
+            {
+                "#include \"Packages/com.unity.render-pipelines.high-definition/Runtime/RenderPipeline/ShaderPass/ShaderPassDepthOnly.hlsl\"",
+            },
+            PixelShaderSlots = new List<int>()
+            {
+                HDLitMasterNode.AlphaSlotId,
+                HDLitMasterNode.AlphaThresholdDepthPostpassSlotId,
+                HDLitMasterNode.DepthOffsetSlotId,
+            },
+            VertexShaderSlots = new List<int>()
+            {
+                HDLitMasterNode.PositionSlotId,
+                HDLitMasterNode.VertexNormalSlotID,
+                HDLitMasterNode.VertexTangentSlotID
+            },
+            UseInPreview = false,
         };
 
         Pass m_PassRaytracingIndirect = new Pass()
@@ -1323,6 +1347,11 @@ namespace UnityEditor.Rendering.HighDefinition
                 if (transparentDepthPostpassActive)
                 {
                     GenerateShaderPassLit(masterNode, m_PassTransparentDepthPostpass, mode, subShader, sourceAssetDependencyPaths);
+                }
+
+                if (masterNode.rayTracing.isOn)
+                {
+                    GenerateShaderPassLit(masterNode, m_PassRayTracingPrepass, mode, subShader, sourceAssetDependencyPaths);
                 }
             }
             subShader.Deindent();